+++ docs/examples/webhooks/delete.php
@@ -1,0 +1,5 @@
+<?php
+
+use Luigel\Paymongo\Facades\Paymongo;
+
+Paymongo::webhooks()->delete('hook_9VrvpRkkYqK6twbhuvcVTtjM'); // true, or throws